moving styles into chatPageStyles
