+++ pong_frontend/src/components/div/channel_buttons_div.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import  {ChatProps} from '../../interfaces/channel.interface';
 import { addAdminPopUp, banUserPopUp, changePasswordPopUp, kickUserPopUp, muteUserPopUp } from './channel_popups';
 import { deleteChannel } from '../../api/channel/channel.api';
-import { ChannelInfo } from './chat_utils';
+import { ChannelInfo } from '../mainPages/ChatPageStyles';
 
 export interface LoadingProps {
 	loadingChannelPanel: boolean;
